COMP: Provide basic completion via completion contributor
