+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsAssocTypeBindingReferenceImpl.kt
@@ -7,10 +7,7 @@ package org.rust.lang.core.resolve.ref
 
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsAssocTypeBinding
-import org.rust.lang.core.psi.RsTraitItem
-import org.rust.lang.core.psi.RsTypeAlias
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processAssocTypeVariants
 
@@ -22,7 +19,4 @@ class RsAssocTypeBindingReferenceImpl(
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processAssocTypeVariants(element, it) }
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processAssocTypeVariants(element, it) }
 }
