+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsExternCrateReferenceImpl.kt
@@ -8,7 +8,6 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsExternCrateItem
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processExternCrateResolveVariants
 
@@ -18,9 +17,6 @@ class RsExternCrateReferenceImpl(
     RsReference {
 
     override val RsExternCrateItem.referenceAnchor: PsiElement get() = referenceNameElement
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processExternCrateResolveVariants(element, true, it) }
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.name!!) { processExternCrateResolveVariants(element, false, it) }
