+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsLabelReferenceImpl.kt
@@ -8,7 +8,6 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsLabel
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processLabelResolveVariants
 
@@ -23,7 +22,4 @@ class RsLabelReferenceImpl(
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processLabelResolveVariants(element, it) }
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processLabelResolveVariants(element, it) }
 }
