+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsLifetimeReferenceImpl.kt
@@ -8,7 +8,6 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsLifetime
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processLifetimeResolveVariants
 
@@ -23,7 +22,4 @@ class RsLifetimeReferenceImpl(
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processLifetimeResolveVariants(element, it) }
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processLifetimeResolveVariants(element, it) }
 }
