+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsModReferenceImpl.kt
@@ -8,7 +8,6 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsModDeclItem
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processModDeclResolveVariants
 
@@ -18,9 +17,6 @@ class RsModReferenceImpl(
     RsReference {
 
     override val RsModDeclItem.referenceAnchor: PsiElement get() = identifier
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processModDeclResolveVariants(element, it) }
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processModDeclResolveVariants(element, it) }
