+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsPatBindingReferenceImpl.kt
@@ -8,7 +8,6 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsPatBinding
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processPatBindingResolveVariants
 
@@ -22,9 +21,6 @@ class RsPatBindingReferenceImpl(
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processPatBindingResolveVariants(element, false, it) }
-
-    override fun getVariants(): Array<out Any> =
-        collectCompletionVariants { processPatBindingResolveVariants(element, true, it) }
 
     override fun isReferenceTo(element: PsiElement): Boolean {
         val target = resolve()
