+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsPathReferenceImpl.kt
@@ -5,7 +5,6 @@
 
 package org.rust.lang.core.resolve.ref
 
-import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.psi.PsiElement
 import com.intellij.psi.ResolveResult
 import org.rust.lang.core.psi.*
@@ -13,7 +12,6 @@ import org.rust.lang.core.psi.ext.*
 import org.rust.lang.core.resolve.*
 import org.rust.lang.core.types.BoundElement
 import org.rust.lang.core.types.Substitution
-import org.rust.lang.core.types.infer.containsTyOfClass
 import org.rust.lang.core.types.infer.foldTyInferWith
 import org.rust.lang.core.types.infer.resolve
 import org.rust.lang.core.types.infer.substitute
@@ -30,24 +28,6 @@ class RsPathReferenceImpl(
     RsPathReference {
 
     override val RsPath.referenceAnchor: PsiElement get() = referenceNameElement
-
-    override fun getVariants(): Array<out LookupElement> {
-        val lookup = ImplLookup.relativeTo(element)
-        return collectCompletionVariants {
-            processPathResolveVariants(
-                lookup,
-                element,
-                true,
-                filterAssocTypes(
-                    element,
-                    filterCompletionVariantsByVisibility(
-                        filterPathCompletionVariantsByTraitBounds(it, lookup),
-                        element.containingMod
-                    )
-                )
-            )
-        }
-    }
 
     override fun isReferenceTo(element: PsiElement): Boolean {
         val target = resolve()
@@ -81,43 +61,6 @@ class RsPathReferenceImpl(
         override fun invoke(element: RsPath): List<BoundElement<RsElement>> {
             return resolvePath(element)
         }
-    }
-}
-
-private fun filterAssocTypes(
-    path: RsPath,
-    processor: RsResolveProcessor
-): RsResolveProcessor {
-    val qualifier = path.path
-    val allAssocItemsAllowed =
-        qualifier == null || qualifier.hasCself || qualifier.reference.resolve() is RsTypeParameter
-    return if (allAssocItemsAllowed) processor else fun(it: ScopeEntry): Boolean {
-        if (it is AssocItemScopeEntry && (it.element is RsTypeAlias)) return false
-        return processor(it)
-    }
-}
-
-private fun filterPathCompletionVariantsByTraitBounds(
-    processor: RsResolveProcessor,
-    lookup: ImplLookup
-): RsResolveProcessor {
-    val cache = mutableMapOf<RsImplItem, Boolean>()
-    return fun(it: ScopeEntry): Boolean {
-        if (it !is AssocItemScopeEntry) return processor(it)
-        if (it.source !is TraitImplSource.ExplicitImpl) return processor(it)
-
-        val receiver = it.subst[TyTypeParameter.self()] ?: return processor(it)
-        // Don't filter partially unknown types
-        if (receiver.containsTyOfClass(TyUnknown::class.java)) return processor(it)
-        // Filter members by trait bounds (try to select all obligations for each impl)
-        // We're caching evaluation results here because we can often complete members
-        // in the same impl and always have the same receiver type
-        val canEvaluate = cache.getOrPut(it.source.value) {
-            lookup.ctx.canEvaluateBounds(it.source.value, receiver)
-        }
-        if (canEvaluate) return processor(it)
-
-        return false
     }
 }
