+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsStructExprFieldReferenceImpl.kt
@@ -10,7 +10,6 @@ import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsPsiFactory
 import org.rust.lang.core.psi.RsStructLiteralField
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.resolve.collectCompletionVariants
 import org.rust.lang.core.resolve.collectResolveVariants
 import org.rust.lang.core.resolve.processStructLiteralFieldResolveVariants
 
@@ -23,8 +22,7 @@ class RsStructLiteralFieldReferenceImpl(
 
     override val cacheDependency: ResolveCacheDependency get() = ResolveCacheDependency.LOCAL_AND_RUST_STRUCTURE
 
-    override fun getVariants(): Array<out LookupElement> =
-        collectCompletionVariants { processStructLiteralFieldResolveVariants(element, true, it) }
+    override fun getVariants(): Array<out LookupElement> = LookupElement.EMPTY_ARRAY
 
     override fun resolveInner(): List<RsElement> =
         collectResolveVariants(element.referenceName) { processStructLiteralFieldResolveVariants(element, false, it) }
